guide: add note about changing mutex rate during execution
When a mutex profile is reported, the counts and delays in the profile
are scaled based on the most recently set rate. So if an application has
changed its profile rate at runtime and collected different samples at
different sampling rates, the samples won't be adjusted correctly when
the pprof is written.