guide: warn about block profiler overhead (#15)
After internal investigations, we found that poor block profiler
configuration can lead to noticeable, and in some cases quite
significant, overhead in production. Add a warning. I also updated the
recommended rates to a 100,000,000, which is a more conservative rate
and would hopefully guard against unexpected performance issues if the
suggested configuration was applied to a production app.

Co-authored-by: Nick Ripley <nick.ripley@datadoghq.com>